+++ src/components/radio/index.ts
@@ -1,0 +1,2 @@
+export { default as Radio } from "./Radio";
+export { default as RadioGroup } from "./RadioGroup";
